+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added a workaround for plugins that improperly defer part of their
+  initialization process without telling the host. This fixes the Roland Cloud
+  plugins.
+
 ## [1.1.2] - 2020-05-09
 
 ### Fixed